src/test/java/mariangela/mar15/ProblemsTest.java
package mariangela.mar15;

import static org.hamcrest.core.Is.*;
import static org.junit.Assert.*;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import org.junit.Test;


public class ProblemsTest {

	@Test
	public void testKillDuplicates() {
		
		List<Integer> array = Arrays.asList(1, 1, 1, 4, 5, 6, 4, 6, 5);
		
		List<Integer> results = Problems.killDuplicates(array);
		assertThat(results.size(), is(4));
		assertTrue(results.contains(1));
		assertTrue(results.contains(4));
		assertTrue(results.contains(5));
		assertTrue(results.contains(6));
	}

}
